application/requirements.txt
aiotask-context==0.6.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
idna==2.10
logfmt==0.4
Pillow==8.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyjwt-rsa==0.3.9
python-dateutil==2.8.1
PyYAML==5.3.1
qrcode==6.1
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
tornado==6.0.4
urllib3==1.25.10
alembic==1.4.3
requests==2.24.0
